+++ core/app/[locale]/(default)/account/addresses/page.tsx
@@ -110,7 +110,7 @@ export default async function Addresses({ params, searchParams }: Props) {
       createLabel={t('create')}
       deleteLabel={t('delete')}
       editLabel={t('edit')}
-      fields={[...fields, { name: 'id', type: 'hidden' }]}
+      fields={[...fields, { name: 'id', type: 'hidden', label: 'ID' }]}
       setDefaultLabel={t('setDefault')}
       showAddFormLabel={t('cta')}
       title={t('title')}
