+++ core/vibes/soul/form/dynamic-form/index.tsx
@@ -24,7 +24,7 @@ import { FormStatus } from '@/vibes/soul/form/form-status';
 import { Input } from '@/vibes/soul/form/input';
 import { NumberInput } from '@/vibes/soul/form/number-input';
 import { RadioGroup } from '@/vibes/soul/form/radio-group';
-import { Select } from '@/vibes/soul/form/select';
+import { SelectField } from '@/vibes/soul/form/select-field';
 import { SwatchRadioGroup } from '@/vibes/soul/form/swatch-radio-group';
 import { Textarea } from '@/vibes/soul/form/textarea';
 import { Button, ButtonProps } from '@/vibes/soul/primitives/button';
@@ -263,7 +263,7 @@ function DynamicFormField({
 
     case 'select':
       return (
-        <Select
+        <SelectField
           errors={formField.errors}
           key={field.name}
           label={field.label}
