+++ core/vibes/soul/sections/cart/shipping-form/index.tsx
@@ -16,7 +16,7 @@ import { FormStatus } from '@/vibes/soul/form/form-status';
 import { Input } from '@/vibes/soul/form/input';
 import { Label } from '@/vibes/soul/form/label';
 import { RadioGroup } from '@/vibes/soul/form/radio-group';
-import { Select } from '@/vibes/soul/form/select';
+import { SelectField } from '@/vibes/soul/form/select-field';
 import { Button } from '@/vibes/soul/primitives/button';
 
 import { shippingActionFormDataSchema } from '../schema';
@@ -267,7 +267,7 @@ export function ShippingForm({
           className={clsx('mt-4 space-y-4', { hidden: !showAddressForm })}
         >
           {Array.isArray(countries) ? (
-            <Select
+            <SelectField
               errors={addressFields.country.errors}
               key={addressFields.country.id}
               label={countryLabel}
@@ -277,7 +277,7 @@ export function ShippingForm({
               onValueChange={countryControl.change}
               options={countries}
               placeholder=""
-              value={countryControl.value}
+              value={countryControl.value ?? ''}
             />
           ) : (
             <Input
@@ -295,7 +295,7 @@ export function ShippingForm({
           />
           <div className="flex gap-3">
             {Array.isArray(states) ? (
-              <Select
+              <SelectField
                 disabled={addressFields.country.value === undefined}
                 errors={addressFields.state.errors}
                 key={addressFields.state.id}
@@ -308,7 +308,7 @@ export function ShippingForm({
                   states.find((s) => s.country === addressFields.country.value)?.states ?? []
                 }
                 placeholder=""
-                value={stateControl.value}
+                value={stateControl.value ?? ''}
               />
             ) : (
               <Input
