+++ core/vibes/soul/sections/product-detail/product-detail-form.tsx
@@ -23,7 +23,7 @@ import { FormStatus } from '@/vibes/soul/form/form-status';
 import { Input } from '@/vibes/soul/form/input';
 import { NumberInput } from '@/vibes/soul/form/number-input';
 import { RadioGroup } from '@/vibes/soul/form/radio-group';
-import { Select } from '@/vibes/soul/form/select';
+import { SelectField } from '@/vibes/soul/form/select-field';
 import { SwatchRadioGroup } from '@/vibes/soul/form/swatch-radio-group';
 import { Textarea } from '@/vibes/soul/form/textarea';
 import { Button } from '@/vibes/soul/primitives/button';
@@ -318,7 +318,7 @@ function FormField({
 
     case 'select':
       return (
-        <Select
+        <SelectField
           errors={formField.errors}
           key={formField.id}
           label={field.label}
